move fns to paths
